% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_fpc.r
\name{calc_fpc}
\alias{calc_fpc}
\title{Calculates point(s) of funnel curve(s)}
\usage{
calc_fpc(
  ci_or_alpha,
  probability,
  sample_size,
  upper_or_lower = 0,
  trunc_val = 1
)
}
\arguments{
\item{ci_or_alpha}{confidence interval (if input >1), alpha value (if input <1)}

\item{probability}{probability of the result; this is centre of the funnel plot (y value)}

\item{sample_size}{size of the sample (x value)}

\item{upper_or_lower}{determines the output; 0: returns upper and lower, 1: returns lower only, 2: returns upper only}

\item{trunc_val}{truncate nonsensical values; default is: return only values between 0-1 and NA for values outside 0-1}
}
\value{
output value(s) between 0-1 (unless trunc_val is not 1)
}
\description{
Returns the 'y' values(s) of the funnel curve(s) for one 'x' value
}
\details{
This function calculates the 'y' value(s) of (a) funnel curve(s) for a given confidence interval (or alpha value), 'x' value (sample size), and funnel centre (probability). Options include: output only the upper or lower value, and truncate the results, if they are not realistic.
}
\examples{
calc_fpc(95, 0.269, 100)
calc_fpc(95, 0.269, 100, 2)
calc_fpc(99, 0.05, 100)

}
\keyword{internal}
