% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controlchart2cmpproviders.r
\name{controlchart2cmpproviders}
\alias{controlchart2cmpproviders}
\title{Creates a control chart with an arbitrary number of control lines}
\usage{
controlchart2cmpproviders(
  x_var,
  y_var,
  labels = NULL,
  limits = c(95, 99.9),
  control_chart_centre_line = NULL,
  addlabels = "NONE",
  labels_to_display = NULL,
  x_label = NULL,
  y_label = NULL,
  plot_title = NULL,
  show_legend = TRUE,
  y_percent = FALSE,
  sort_by_y_norm = TRUE,
  showoutlier_arr = FALSE
)
}
\arguments{
\item{x_var}{x values on funnel plot, e.g. number of specimens}

\item{y_var}{y values on funnel plot, e.g. diagnostic rate (0-1)}

\item{labels}{labels that describe the (x,y) data points, e.g. name of institution or provider identification; default: the data points are numbered (if undefined)}

\item{limits}{limits (confidence intervals); default: 95\%, 99.9\%}

\item{control_chart_centre_line}{defines the control chart centre line; default: median of y_var}

\item{addlabels}{whether to display labels on the funnel plot; accepted values are:
\itemize{
\item{\code{"none"}}{ : do not add labels (default)}
\item{\code{"all"}}{ : add labels to all points}
\item{\code{"outliers"}}{ : label only outliers outside of the inner funnels}
\item{\code{"outliers2"}}{ : label only outliers outside of second innermost funnel}
\item{\code{"only_oaf"}}{ : label only outliers outside of all funnels}
\item{\code{"list"}}{ : display the list of labels given by argument \code{labels_to_display}}
}}

\item{labels_to_display}{selected labels to display (see argument \emph{addlabels})}

\item{x_label}{x axis label}

\item{y_label}{y axis label}

\item{plot_title}{title of the (funnel) plot}

\item{show_legend}{TRUE shows legend at bottom of plot}

\item{y_percent}{y axis in percent (multiplies y_var_normed by 100)}

\item{sort_by_y_norm}{sort by the normed y value}

\item{showoutlier_arr}{show the outlier array}
}
\value{
funnel plot object
}
\description{
Control charts to compare (healthcare) providers or institutions
}
\details{
A function to compare provider rate data.  The implementation is based on Bonert et al. 
(2023) https://doi.org/10.1038/s41598-022-26962-w
}
\examples{
library(COUNT)
library(dplyr)
data(medpar)

print("Number of cases by provider ...")
provider_count=as.data.frame(table(medpar$provnum))
colnames(provider_count) = c('PROVIDER', 'COUNT')
provider_count

print("Purging providers with less than 20 cases ...")
requiredcases=20
medpar_trunc <- medpar[medpar$provnum \%in\% provider_count[provider_count$COUNT>=requiredcases,]$PROVIDER,]

print("Number of cases by provider in purged data set ...")
provider_count_trunc=as.data.frame(table(medpar_trunc$provnum))
colnames(provider_count_trunc) = c('PROVIDER', 'COUNT')
provider_count_trunc

medpar_trunc_died <- medpar_trunc[medpar_trunc$died==1,]

print("Number of died by provider in purged data set ...")
provider_count_trunc_died=as.data.frame(table(medpar_trunc_died$provnum))
colnames(provider_count_trunc_died) = c('PROVIDER', 'COUNT')
provider_count_trunc_died

print("Death rate by provider ...")
y_var = provider_count_trunc_died[,2]/ provider_count_trunc[,2]
y_var

x_var = provider_count_trunc[,2]
provider_labels = provider_count_trunc[,1]

print("Creating control charts ...")
cc1=controlchart2cmpproviders(x_var, y_var, limits=c(95,99.9,99.9999), addlabels = 1, 
  x_label = "Provider", y_label = "Normalized Death Rate by Provider", 
  plot_title = "Normalized Death Rate by Provider") 
cc2=controlchart2cmpproviders(x_var, y_var, limits=c(95,99.9,99.9999), labels=provider_labels, 
  addlabels = "OUTLIERS", x_label = "Provider", y_label = "Normalized Provider Death Rate", 
  plot_title = "Normalized Death Rate by Provider with Outliers Labelled", do_not_sort_by_y_norm=1)
cc3=controlchart2cmpproviders(x_var, y_var, limits=c(95,99.9,99.9999), labels=provider_labels, 
  addlabels = "OUTLIERS", x_label = "Provider", y_label = "Normalized Provider Death Rate", 
  plot_title = "Normalized Death Rate by Provider with Outliers Labelled")

}
