% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funnel2cmpproviders.r
\name{funnel2cmpproviders}
\alias{funnel2cmpproviders}
\title{Creates a funnel plot with an arbitrary number of funnels}
\usage{
funnel2cmpproviders(
  x_var,
  y_var,
  labels = NULL,
  limits = c(95, 99.9),
  funnel_centre_line = NULL,
  addlabels = "NONE",
  labels_to_display = NULL,
  x_label = NULL,
  y_label = NULL,
  plot_title = NULL,
  x_lower_limit = NULL,
  x_upper_limit = NULL,
  y_lower_limit = NULL,
  y_upper_limit = NULL,
  x_fpad_upper = 0.1,
  x_fpad_lower = 0.03,
  y_fsd = 0.5,
  funnel_curve_calculation_parameter = 3,
  fc_min_dx = 0.5,
  fc_max_y_err = 0.003,
  x2plot = NULL,
  show_legend = TRUE,
  y_percent = FALSE
)
}
\arguments{
\item{x_var}{x values on funnel plot, e.g. number of specimens}

\item{y_var}{y values on funnel plot, e.g. diagnostic rate (0-1)}

\item{labels}{labels that describe the (x,y) data points, e.g. name of institution or provider identification; default: the data points are numbered (if undefined)}

\item{limits}{limits (confidence intervals); default: 95\%, 99.9\%}

\item{funnel_centre_line}{defines the funnel centre line; default: median of y_var}

\item{addlabels}{whether to display labels on the funnel plot; accepted values are:
\itemize{
\item{\code{"none"}}{ : do not add labels (default)}
\item{\code{"all"}}{ : add labels to all points}
\item{\code{"outliers"}}{ : label only outliers outside of the inner funnels}
\item{\code{"outliers2"}}{ : label only outliers outside of second innermost funnel}
\item{\code{"only_oaf"}}{ : label only outliers outside of all funnels}
\item{\code{"list"}}{ : display the list of labels given by argument \code{labels_to_display}}
}}

\item{labels_to_display}{selected labels to display (see argument \emph{addlabels})}

\item{x_label}{x axis label}

\item{y_label}{y axis label}

\item{plot_title}{title of the (funnel) plot}

\item{x_lower_limit}{lower x limit (min x value on plot)}

\item{x_upper_limit}{upper x limit (max x value on plot)}

\item{y_lower_limit}{lower y limit (min y value on plot)}

\item{y_upper_limit}{upper y limit (max y value on plot)}

\item{x_fpad_upper}{upper padding as fraction of x axis range}

\item{x_fpad_lower}{lower padding as fraction of x axis range}

\item{y_fsd}{padding as fraction of standard deviation of y values}

\item{funnel_curve_calculation_parameter}{method used to calculate the funnel curves / determine the x values used to calculate the funnel curves; 1: uses geom_function to calculate/plot funnel curves, 2 = uses geom_line to plot funnel curves, 3: uses geom_line to plot funnel curves, calculates ideal x using \emph{get_ideal_x.r} (default)}

\item{fc_min_dx}{funnel curve parameter, minimum delta x}

\item{fc_max_y_err}{funnel curve parameter, maximum y error}

\item{x2plot}{vector defines extra x values used to calculate the funnel curves}

\item{show_legend}{TRUE shows legend at bottom of plot}

\item{y_percent}{y axis in percent (multiplies y_var by 100)}
}
\value{
funnel plot object
}
\description{
Funnel plots to compare (healthcare) providers or institutions
}
\details{
A function to compare provider rate data.  The implementation is based on Bonert et al. 
(2017) https://doi.org/10.4103/jpi.jpi_50_17
}
\examples{
library(COUNT)
library(dplyr)
data(medpar)

print("Number of cases by provider ...")
provider_count=as.data.frame(table(medpar$provnum))
colnames(provider_count) = c('PROVIDER', 'COUNT')
provider_count

print("Purging providers with less than 20 cases ...")
requiredcases=20
medpar_trunc <- medpar[medpar$provnum \%in\% provider_count[provider_count$COUNT>=requiredcases,]$PROVIDER,]

print("Number of cases by provider in purged data set ...")
provider_count_trunc=as.data.frame(table(medpar_trunc$provnum))
colnames(provider_count_trunc) = c('PROVIDER', 'COUNT')
provider_count_trunc

medpar_trunc_died <- medpar_trunc[medpar_trunc$died==1,]

print("Number of died by provider in purged data set ...")
provider_count_trunc_died=as.data.frame(table(medpar_trunc_died$provnum))
colnames(provider_count_trunc_died) = c('PROVIDER', 'COUNT')
provider_count_trunc_died

print("Death rate by provider ...")
y_var = provider_count_trunc_died[,2]/ provider_count_trunc[,2]
y_var

x_var = provider_count_trunc[,2]
provider_labels = provider_count_trunc[,1]

print("Creating funnel plots ...")
fp1=funnel2cmpproviders(x_var, y_var, limits=c(95,99.9,99.9999), addlabels = 1, 
  x_label = "Provider \crVolume (Patients Cared for by Provider)", y_label = "Death Rate (Deaths by 
  Provider/Patients Cared for by Provider)", plot_title = "Death Rate by Provider and Volume") 
fp2=funnel2cmpproviders(x_var, y_var, limits=c(95,99.9,99.9999), labels=provider_labels, 
  addlabels = "OUTLIERS", x_label = "Provider Volume (Patients Cared for by Provider)", 
  y_label = "Provider Death Rate (Deaths by Provider/Patients Cared for by Provider)", 
  plot_title = "Death Rate by Provider and Volume with Outliers Labelled")

}
