% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ideal_x.r
\name{get_ideal_x}
\alias{get_ideal_x}
\title{Calculates ideal x values to obtain a smooth funnel curve}
\usage{
get_ideal_x(
  x_lower_limit,
  x_upper_limit,
  y_lower_limit,
  y_upper_limit,
  limits,
  funnel_centre_line,
  fc_min_dx,
  fc_max_y_err
)
}
\arguments{
\item{x_lower_limit}{minimum x value for plot (0-1)}

\item{x_upper_limit}{maximum x value for plot (0-1)}

\item{y_lower_limit}{minimum y value for plot (0-1)}

\item{y_upper_limit}{maximum y value for plot (0-1)}

\item{limits}{funnel plot limits, e.g. 95 or 99.9}

\item{fc_min_dx}{minimum delta x}

\item{fc_max_y_err}{maximal y error}
}
\value{
a vector that represent the ideal x values to obtain a smooth funnel curve
}
\description{
Returns the x values for the funnel curve
}
\details{
This function calculates the x values to obtain a smooth funnel curve. Inputs are the plot limits, confidence intervals, funnel centre line and two control paramters ("fc_min_dx" and "fc_max_y_err"). The function approximates a y value between two x values by linear interpolation and then calculates the error in y using the function "calc_fpc". If the error in y exceeds "fc_max_y_err" and distance between the x values is greater than "fc_min_dx": an x value is added.
}
\examples{
get_ideal_x(20, 100, .1, .9, c(95,99), 0.45, 0.5, 0.003)

}
\keyword{internal}
