% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_r_normed.r
\name{get_r_normed}
\alias{get_r_normed}
\title{Calculates the normalized rate value}
\usage{
get_r_normed(r_measured, r_ideal, n_normed, n_measured)
}
\arguments{
\item{r_measured}{measured rate (0-1)}

\item{r_ideal}{ideal measure (0-1)}

\item{n_normed}{normalized n}

\item{n_measured}{measured n}
}
\value{
the normalized rate
}
\description{
Returns normalized rate value
}
\details{
This function calculates the normed r. Inputs are the measured rate, ideal rate, normalized count and the measured count.  
The implementation is based on Bonert et al. (2023) https://doi.org/10.1038/s41598-022-26962-w
}
\examples{
get_r_normed(0.14, 0.10, 198, 82)
get_r_normed(0.0, 0.10, 198, 82)

}
\keyword{internal}
